+++ src/steps/index.ts
@@ -1,6 +1,14 @@
 import { accountSteps } from './account';
-import { accessSteps } from './access';
-
-const integrationSteps = [...accountSteps, ...accessSteps];
+import { agentSteps } from './agent';
+import { groupSteps } from './agent-groups';
+import { roleSteps } from './agent-roles';
+import { ticketSteps } from './tickets';
+const integrationSteps = [
+  ...accountSteps,
+  ...agentSteps,
+  ...groupSteps,
+  ...roleSteps,
+  ...ticketSteps,
+];
 
 export { integrationSteps };